+++ ConfigManager.cs
@@ -12,6 +12,10 @@ public class ConfigManager {
 
     public static BoolField Enable;
     public static ColorField AccentColor;
+    public static BoolField EnableTrackers;
+    public static IntField TrackerThreshold;
+    public static BoolField TrackerShowEnemyNames;
+    public static ColorField TrackerColor;
 
     public static FloatField StartScale;
     public static FloatField EndScale;
@@ -20,6 +24,9 @@ public class ConfigManager {
     public static FloatField ScalingDuration;
     public static FloatField ScalingDelay;
     public static FloatField VisibleDuration;
+
+    public static FloatField TrackerScale;
+    public static FloatSliderField TrackerMarkOpacity;
 
     public static void Initialize() {
     }
@@ -30,20 +37,30 @@ public class ConfigManager {
         if (File.Exists(iconPath)) config.SetIconWithURL(iconPath);
 
         new ConfigHeader(config.rootPanel, "", 10);
-        new ConfigHeader(config.rootPanel, "-- BASIC --", 20);
+        new ConfigHeader(config.rootPanel, "-- CROSSES --", 22);
         Enable = new BoolField(config.rootPanel, "Enable death crosses", "enable", true);
         AccentColor = new ColorField(config.rootPanel, "Cross accent color", "accentColor", new Color(219f / 255f, 30f / 255f, 57f / 255f));
+
+        new ConfigHeader(config.rootPanel, "", 10);
+        new ConfigHeader(config.rootPanel, "-- ENEMY TRACKERS --", 22);
+        EnableTrackers = new BoolField(config.rootPanel, "Enable trackers", "enableTrackers", false);
+        TrackerThreshold = new IntField(config.rootPanel, "Reveal last remaining x enemies", "trackerThreshold", 5);
+        TrackerShowEnemyNames = new BoolField(config.rootPanel, "Show enemy names", "trackerShowEnemyNames", true);
+        TrackerColor = new ColorField(config.rootPanel, "Enemy tracker mark color", "trackerColor", new Color(241f / 255f, 182f / 255f, 19f / 255f));
 
         new ConfigHeader(config.rootPanel, "", 10);
         new ConfigHeader(config.rootPanel, "-- ADVANCED --", 20);
         StartScale = new FloatField(config.rootPanel, "Cross starting scale", "startScale", 2f);
         EndScale = new FloatField(config.rootPanel, "Cross ending scale", "endScale", 0.35f);
-        EnemyHealthThreshold = new FloatField(config.rootPanel, "Enemy health threshold", "enemyHealthThreshold", 0);
+        EnemyHealthThreshold = new FloatField(config.rootPanel, "Cross enemy health threshold", "enemyHealthThreshold", 0);
         CrossMarkOpacity = new FloatSliderField(config.rootPanel, "Cross mark opacity", "crossMarkOpacity",
             new Tuple<float, float>(0f, 1f), 1f);
         ScalingDelay = new FloatField(config.rootPanel, "Scaling delay", "scalingDelay", 0.000f);
         ScalingDuration = new FloatField(config.rootPanel, "Scaling duration", "scalingDuration", 0.1333f);
         VisibleDuration = new FloatField(config.rootPanel, "Visible duration", "visibleDuration", 0.5333f);
 
+        TrackerScale = new FloatField(config.rootPanel, "Tracker mark scale", "trackerScale", 1f);
+        TrackerMarkOpacity = new FloatSliderField(config.rootPanel, "Tracker mark opacity", "trackerMarkOpacity",
+            new Tuple<float, float>(0f, 1f), 1f);
     }
 }