+++ EnemyListener.cs
@@ -1,0 +1,10 @@
+﻿namespace Crossover;
+
+public static class EnemyListener {
+    public static event OnSomeEnemyDied SomeEnemyDied;
+    public delegate void OnSomeEnemyDied();
+
+    internal static void NotifyDeath() {
+        SomeEnemyDied?.Invoke();
+    }
+}