+++ Patches/EnemyPatch.cs
@@ -1,17 +1,27 @@
 ﻿using HarmonyLib;
+using Notiffy.API;
 using UnityEngine;
 
 namespace Crossover.Patches;
 
 [HarmonyPatch(typeof(EnemyIdentifier))]
-class EnemyPatch {
+internal class EnemyPatch {
     [HarmonyPostfix]
     [HarmonyPatch("ProcessDeath")]
-    static void Postfix(EnemyIdentifier __instance) {
+    static void EnemyDied(EnemyIdentifier __instance) {
         if (!ConfigManager.Enable.value || __instance.health < ConfigManager.EnemyHealthThreshold.value) return;
-        GameObject cross = Object.Instantiate(Plugin.DeathCrossPrefab, Plugin.DeathCrossCanvas.transform);
+        GameObject cross = Object.Instantiate(Plugin.DeathCrossPrefab, Plugin.CrossoverCanvas.transform);
         CrossBehavior crossBehavior = cross.AddComponent<CrossBehavior>();
-        crossBehavior.enemy = __instance;
+        crossBehavior.SetEnemy(__instance);
         cross.SetActive(true);
+        EnemyListener.NotifyDeath();
+    }
+
+    [HarmonyPostfix]
+    [HarmonyPatch("Start")]
+    static void EnemySpawned(EnemyIdentifier __instance) {
+        if (!ConfigManager.EnableTrackers.value) return;
+        EnemyIndicatorController controller = __instance.gameObject.AddComponent<EnemyIndicatorController>();
+        controller.SetEnemy(__instance);
     }
 }