+++ package/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.1
 
+- Added indicator for the last few remaining enemies
 - Made cross durations customizable
 - Internal: create crosses when enemies die instead of when they spawn. Should help performance a bit.